Move corpus encode/decode to importable internal/corpus package
Move Decode, Encode, and Generate into internal/corpus so both
internal/cmd/fuzz and internal/cmd/seedcorpus can import it.
